ai-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
zhipuai>=2.0.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
